backend/requirements.txt
gunicorn==22.0.0

# FastAPI and Server
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
websockets==12.0

# Database
sqlalchemy==2.0.31
psycopg2-binary==2.9.10
alembic==1.16.5

# Data and Background Tasks
celery[redis]==5.5.3
pandas==2.2.2
redis==5.0.7

# Authentication and Security
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
PyJWT==2.10.1
google-auth==2.32.0
google-auth-oauthlib==1.2.2
requests==2.32.3

# Pydantic and Validation
pydantic==2.8.2
email-validator==2.3.0
pytz==2024.1

# Environment and Email
python-dotenv==1.0.1
fastapi-mail==1.4.1
slowapi==0.1.9
cryptography==46.0.2

google-generativeai==0.7.2
markdown2==2.5.3